se han hechoalgunas cosas

+++ Ejercicios/ActividadesU4/src/com/Ejercicio11.java
@@ -5,23 +5,23 @@ public class Ejercicio11 {
 	//MAIN
 	public static void main(String[] args) {
 		//Definimos los parametros que vamos a usar
-		int a = 3;
+		double a = 3;
 		int n = 5;
 		
 		//ejecutamos la funcion iterativa (con bucle) con a y n como parametros
-		int resultado_iterativ = iterativa(a, n);
+		double resultado_iterativ = iterativa(a, n);
 		System.out.println("Iterativa = "+resultado_iterativ);
 		
 		//ejecutamos la funcion recursiva (una funcion que llama a otra funcion)
-		int resutlado_recursiv = recursiva(a,n);
+		double resutlado_recursiv = recursiva(a,n);
 		System.out.println("Recursiva = "+ resutlado_recursiv);
 	}
 	
 	//FUNCIONES-------------
 	
 	//ITERATIVA
-	public static int iterativa(int a, int n) {
-		int result = a; //acumulador
+	public static double iterativa(double a, int n) {
+		double result = a; //acumulador
 		n--; //le quitamos 1 a n como para hacer el primer a * a
 		
 		//en el bucle le damos al acumulador hasta que n sea 0
@@ -33,7 +33,7 @@ public class Ejercicio11 {
 	}
 	
 	//RECURSIVA
-	public static int recursiva(int a, int n) {
+	public static double recursiva(double a, int n) {
 		
 		if (n == 0) {
 			return 1;
